Move `nullSafeConversion` to ExprMapper
