+++ spark-cypher/src/main/scala/org/opencypher/spark/impl/SparkSQLExprMapper.scala
@@ -26,15 +26,15 @@
  */
 package org.opencypher.spark.impl
 
+import org.apache.spark.sql.catalyst.expressions.CaseWhen
 import org.apache.spark.sql.functions.{array_contains => _, translate => _, _}
 import org.apache.spark.sql.types._
-import org.apache.spark.sql.{Column, DataFrame, functions}
+import org.apache.spark.sql.{Column, DataFrame}
 import org.opencypher.okapi.api.types._
 import org.opencypher.okapi.api.value.CypherValue.CypherMap
 import org.opencypher.okapi.impl.exception._
 import org.opencypher.okapi.impl.temporal.TemporalTypesHelper._
 import org.opencypher.okapi.impl.temporal.{Duration => DurationValue}
-import org.opencypher.okapi.ir.api.PropertyKey
 import org.opencypher.okapi.ir.api.expr._
 import org.opencypher.okapi.relational.impl.table.RecordHeader
 import org.opencypher.spark.impl.CAPSFunctions._
@@ -51,6 +51,32 @@ object SparkSQLExprMapper {
   implicit class RichExpression(expr: Expr) {
 
     /**
+      * Converts `expr` with the `withConvertedChildren` function, which is passed the converted child expressions as its
+      * argument.
+      *
+      * Iff the expression has `expr.nullInNullOut == true`, then any child being mapped to `null` will also result in
+      * the parent expression being mapped to null.
+      *
+      * For these expressions the `withConvertedChildren` function is guaranteed to not receive any `null`
+      * values from the evaluated children.
+      */
+    def nullSafeConversion(expr: Expr)(withConvertedChildren: Seq[Column] => Column)
+      (implicit header: RecordHeader, df: DataFrame, parameters: CypherMap): Column = {
+      if (expr.cypherType == CTNull) {
+        NULL_LIT
+      } else {
+        val evaluatedArgs = expr.children.map(_.asSparkSQLExpr)
+        val withConvertedChildrenResult = withConvertedChildren(evaluatedArgs).expr
+        if (expr.children.nonEmpty && expr.nullInNullOut && expr.cypherType.isNullable) {
+          val nullPropagationCases = evaluatedArgs.map(_.isNull.expr).zip(Seq.fill(evaluatedArgs.length)(NULL_LIT.expr))
+          new Column(CaseWhen(nullPropagationCases, withConvertedChildrenResult))
+        } else {
+          new Column(withConvertedChildrenResult)
+        }
+      }
+    }
+
+    /**
       * Attempts to create a Spark SQL expression from the CAPS expression.
       *
       * @param header     the header of the CAPSRecords in which the expression should be evaluated.
@@ -63,7 +89,7 @@ object SparkSQLExprMapper {
         // Evaluate based on already present data; no recursion
         case _: Var | _: HasLabel | _: HasType | _: StartNode | _: EndNode => column_for(expr)
         // Evaluate bottom-up
-        case _ => null_safe_conversion(expr)(convert)
+        case _ => nullSafeConversion(expr)(convert)
       }
       header.getColumn(expr) match {
         case None => outCol
